requirements.txt
# Nebula Core (backend)
fastapi
uvicorn[standard]
pydantic
pydantic-settings
PyYAML
psutil
bcrypt
